+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/Stat.java
@@ -9,9 +9,6 @@ import com.robertx22.age_of_exile.database.registry.SlashRegistryType;
 import com.robertx22.age_of_exile.mmorpg.Ref;
 import com.robertx22.age_of_exile.saveclasses.gearitem.gear_bases.Rarity;
 import com.robertx22.age_of_exile.saveclasses.item_classes.tooltips.TooltipStatWithContext;
-import com.robertx22.age_of_exile.saveclasses.unit.StatData;
-import com.robertx22.age_of_exile.saveclasses.unit.Unit;
-import com.robertx22.age_of_exile.uncommon.effectdatas.EffectData;
 import com.robertx22.age_of_exile.uncommon.enumclasses.Elements;
 import com.robertx22.age_of_exile.uncommon.interfaces.IAutoLocDesc;
 import com.robertx22.age_of_exile.uncommon.interfaces.IAutoLocName;
@@ -38,7 +35,7 @@ public abstract class Stat implements IGUID, IAutoLocName, IWeighted, IRarity, I
 
     public boolean isShown = true;
 
-    public IStatEffect statEffect = new EmptyEffect();
+    public IStatEffect statEffect = null;
 
     public float min_val = -1000;
     public float max_val = Integer.MAX_VALUE;
@@ -212,21 +209,4 @@ public abstract class Stat implements IGUID, IAutoLocName, IWeighted, IRarity, I
         }
     }
 
-    static class EmptyEffect implements IStatEffect {
-
-        @Override
-        public EffectSides Side() {
-            return EffectSides.Source;
-        }
-
-        @Override
-        public int GetPriority() {
-            return 0;
-        }
-
-        @Override
-        public EffectData TryModifyEffect(EffectData effect, Unit source, StatData data, Stat stat) {
-            return effect;
-        }
-    }
 }