+++ src/main/java/com/robertx22/age_of_exile/database/registry/Database.java
@@ -22,6 +22,7 @@ import com.robertx22.age_of_exile.database.data.rarities.MobRarity;
 import com.robertx22.age_of_exile.database.data.runes.Rune;
 import com.robertx22.age_of_exile.database.data.runewords.RuneWord;
 import com.robertx22.age_of_exile.database.data.salvage_outputs.SalvageOutput;
+import com.robertx22.age_of_exile.database.data.skill_gem.SpellGem;
 import com.robertx22.age_of_exile.database.data.spell_schools.SpellSchool;
 import com.robertx22.age_of_exile.database.data.spells.components.Spell;
 import com.robertx22.age_of_exile.database.data.stats.Stat;
@@ -180,6 +181,10 @@ public class Database {
         return getRegistry(SlashRegistryType.TIER);
     }
 
+    public static SlashRegistryContainer<SpellGem> SpellGems() {
+        return getRegistry(SlashRegistryType.SPELL);
+    }
+
     public static SlashRegistryContainer<BaseGearType> GearTypes() {
         return getRegistry(SlashRegistryType.GEAR_TYPE);
     }
@@ -317,6 +322,7 @@ public class Database {
         addRegistry(new SlashRegistryContainer<>(SlashRegistryType.SALVAGE_OUTPUT, null).setIsDatapack());
         addRegistry(new SlashRegistryContainer<>(SlashRegistryType.PLAYER_SKILLS, null).setIsDatapack());
         addRegistry(new SlashRegistryContainer<>(SlashRegistryType.BASE_STATS, null).setIsDatapack());
+        addRegistry(new SlashRegistryContainer<>(SlashRegistryType.SPELL_GEM, null).setIsDatapack());
         addRegistry(new SlashRegistryContainer<>(SlashRegistryType.COMPATIBLE_ITEM,
             CompatibleItem.EMPTY).dontErrorIfEmpty()
             .setIsDatapack()